requirements.txt
appdirs==1.4.3
asgiref==3.2.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-elasticsearch-dsl==7.1.1
django-extensions==2.2.9
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==2.0.2
django-reversion==3.0.7
elasticsearch==7.6.0
elasticsearch-dsl==7.1.0
et-xmlfile==1.0.1
filelock==3.0.12
future==0.15.2
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pipenv==2018.11.26
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
rules==2.2
shiboken2==5.15.0
simpleui==4.0
six==1.10.0
sqlparse==0.3.1
tablib==1.1.0
uritemplate==3.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
xadmin==2.0.1
